+++ hfut/__init__.py
@@ -5,7 +5,7 @@
 from __future__ import unicode_literals
 
 __title__ = 'hfut'
-__version__ = '1.4.1'
+__version__ = '1.4.2'
 __author__ = 'erliang'
 __author_email__ = 'dev@erliang.me'
 __url__ = 'https://github.com/er1iang/hfut'