+++ hfut/log.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 
 from logging import Logger, WARNING, StreamHandler, Formatter
 
-__all__ = ['logger', 'unstable', 'log_result_not_found']
+__all__ = ['logger', 'log_result_not_found']
 
 logger = Logger('hfut', level=WARNING)
 
@@ -17,10 +17,5 @@ sh.setFormatter(fmt)
 logger.addHandler(sh)
 
 
-def unstable(func):
-    logger.warning('%s 功能尚不稳定, 建议使用时验证结果的正确性', func.__name__)
-    return func
-
-
 def log_result_not_found(page):
-    logger.error('没有解析到结果, 请检查你的参数是否正确 \n %s', page)
+    logger.warning('没有解析到结果, 请检查你的参数是否正确 \n %s', page)